+++ packages/@valapi/auth/CHANGELOG.md
@@ -1,3 +1,66 @@
+# 5.0.0-alpha.0
+
+**Add**
+
+-   `AuthRequest`
+
+**Change**
+
+-   ~~`AuthClient`~~ **-->** `Auth`
+-   ~~`AuthClient.refresh()`~~ **-->** `Auth.reauthorize()`
+-   ~~`AuthService.fromUrl(TokenUrl)`~~ **-->** `Auth.uriTokenization(url)`
+-   ~~`AuthService.fromToken(token)`~~
+    -   **-->** `Auth.entitlementsTokenization()`
+    -   **-->** `Auth.regionTokenization()`
+-   ~~`Cookie.authorize()`~~ **-->** `Auth.authorize(ignoreCookie = false)`
+-   ~~`User.loginform(username, password)`~~ **-->** `Auth.login(username, password)`
+-   ~~`AuthCore`~~ **-->** `AuthInstance`
+-   ~~`AuthCore.getSubject(token)`~~ **-->** `AuthInstance.subject`
+-   ~~`AuthCore.getExpirationDate(token)`~~ **-->** `AuthInstance.expirationDate`
+-   ~~`AuthCore.authenticationInfo.isMultifactor`~~ **-->** `AuthInstance.isMultifactor`
+
+**Remove**
+
+-   `AuthClient.verify()`
+-   `AuthClient.fromCookie()`
+-   `AuthService`
+-   `Cookie`
+-   `Multifactor`
+-   `User`
+-   `AuthCore.authenticationInfo`
+
+static
+
+-   `AuthClient.fromCookie()`
+-   `AuthClient.fromJSON()`
+-   `AuthCore.fromJSON()`
+-   `AuthCore.Default`
+-   `AuthService.fromJSON()`
+
+### Typescript
+
+**Add**
+
+-   `EntitlementsTokenResponse`
+-   `RegionTokenResponse`
+-   `UserTokenParse`
+-   `AuthPromiseResponse`
+-   `AuthResponse`
+-   `AuthRequestConfig`
+
+**Change**
+
+-   ~~`AuthService.TokenResponse`~~ **-->** `AuthRequestResponse`
+-   ~~`AuthCore.Json`~~ **-->** `UserAuthInfo`
+-   ~~`AuthCore.ClientPlatfrom`~~ **-->** `AuthRequestPlatfrom`
+-   ~~`AuthCore.Config`~~ **-->** `AuthConfig`
+
+**Remove**
+
+-   `AuthCore.JsonRegion`
+-   `AuthCore.JsonAuthenticationInfoMessage`
+-   `AuthCore.JsonAuthenticationInfo`
+
 # 3.2.1-beta
 
 **Add**